backend: prawns registration somewhat working?

+++ src/main/kotlin/mnu/repository/DistrictHouseRepository.kt
@@ -2,9 +2,10 @@ package mnu.repository
 
 import mnu.model.DistrictHouse
 import org.springframework.data.jpa.repository.*
+import java.math.BigInteger
 import java.util.*
 
 interface DistrictHouseRepository : JpaRepository<DistrictHouse, Long> {
     @Query("select dt.id from district_houses dt;", nativeQuery = true)
-    fun getAllIds(): List<Long>
+    fun getAllIds(): List<BigInteger>
 }
